Updated stroke analysis output formatting

+++ basic_results_analysis.py
@@ -1,6 +1,11 @@
 import data_io
 import pandas as pd
 import parameters as param
+import re
+import time
+from tqdm import tqdm
+
+start_time = time.time()
 
 # See files that are downloaded
 # out = glob.glob(str(data_io.LOCAL_OUTPUT/
@@ -8,10 +13,11 @@ import parameters as param
 # np.sort(out)
 
 def get_basic_results(res_name):
-
     res = pd.read_csv(res_name,index_col='Location')
-    center_cols = res.columns[9:]
-    res['BestCenter'] = res[center_cols].idxmax(axis=1)
+    center_cols = res.columns[10:] # only columns of hospital keys
+    # Returns Best Center Key for each location
+    res['BestCenter'] = res[center_cols].idxmax(axis=1) # idxmax: return index of maximum in each row
+    # Extract Center Key and Type from BestCenter column
     res['BestCenterType'] = res['BestCenter'].apply(lambda x: "CSC" if x.split(' ')
                                                     [1] == '(CSC)' else "PSC")
     res['BestCenterKey'] = res['BestCenter'].apply(lambda x: x.split(' ')[0])
@@ -27,7 +33,7 @@ def _get_most_ce_strategy(agout):
 
 def get_variable_stats_from_aggregated_outcome(agout,var='QALY',strategies=None,suffixes=None):
     if strategies is None:
-        # get most cost-effective strategy
+        # get most cost-effective strategy (one strategy for each file)
         most_ce =  _get_most_ce_strategy(agout)
         for_join = agout.loc[var,most_ce].to_frame().T
         suffixes = ['af']
@@ -42,17 +48,20 @@ def get_variable_stats_from_aggregated_outcome(agout,var='QALY',strategies=None,
     strategies_stats = pd.concat(stats_df_l,axis=1)
     return strategies_stats
 
-for pid in range(500,501):
-    print(pid)
+for pid in tqdm(range(537,581)):
     # get after AHA result
     res_name = list(data_io.LOCAL_OUTPUT.glob(f'pid={pid}*_afAHA.csv'))[0]
-    a_res, center_cols = get_basic_results(res_name)
+    a_res, center_cols = get_basic_results(res_name) # res = dataframe w/ best type & keys , center_cols = all host location columns
+
     # get before AHA result
     res_name = list(data_io.LOCAL_OUTPUT.glob(f'pid={pid}*_beAHA.csv'))[0]
     b_res, center_cols = get_basic_results(res_name)
 
+    # Dataframe of before and after best center key/type for every location
     ab_res = b_res[BEST_CENTER_COLS].join(a_res[BEST_CENTER_COLS],
                                            rsuffix='_af',lsuffix='_be')
+
+    # center_cols is the same for both be and af AHA
     # get list of all possible hospitals to go to, (where cell value is not nan)
     # ab_res = ab_res.join(
     #     pd.DataFrame.from_dict({
@@ -72,23 +81,60 @@ for pid in range(500,501):
         data_io.BASIC_ANALYSIS_OUTPUT / res_name.stem.replace('beAHA','changed.csv'))
 
     qaly_stats_l = []
-    for loc_id in changed_res.index:
-        fname = list(data_io.LOCAL_OUTPUT.glob(f'pid={pid}*loc={loc_id}*afAHA*'))[0]
-        af_agout = read_agout(fname)
-        af_most_ce = _get_most_ce_strategy(af_agout)
-        fname = list(data_io.LOCAL_OUTPUT.glob(f'pid={pid}*loc={loc_id}*beAHA*'))[0]
-        be_agout = read_agout(fname)
-        be_most_ce = _get_most_ce_strategy(be_agout)
-        af_qaly_stats = get_variable_stats_from_aggregated_outcome(af_agout,'QALY',
-                            [af_most_ce,be_most_ce.replace('- most C/E','').strip()],
-                            suffixes=("af","be"))
-        af_qaly_stats.index=[loc_id]
-        # fname = list(data_io.LOCAL_OUTPUT.glob(f'pid={pid}*loc={loc_id}*beAHA*'))[0]
-        # be_qaly_stats = get_variable_stats_from_aggregated_outcome(fname,'QALY')
-        # be_qaly_stats.index=[loc_id]
-        # qaly_stats = be_qaly_stats.join(af_qaly_stats,lsuffix='_be',rsuffix='_af')
-        qaly_stats_l.append(af_qaly_stats)
+    if changed_res.shape[0] != 0: # only run if changes exist
+        # Iterate throughout each location that changed which was the best place to go to
+        for loc_id in changed_res.index:
+            fname = list(data_io.LOCAL_OUTPUT.glob(f'pid={pid}*loc={loc_id}_afAHA*'))[0]
+            af_agout = read_agout(fname) # each individual output file
+            # Get the most cost effective strategy - after (af)
+            af_most_ce = _get_most_ce_strategy(af_agout)
 
-    changed_res = changed_res.join(pd.concat(qaly_stats_l),how='left')
-    changed_res.to_csv(
-        data_io.BASIC_ANALYSIS_OUTPUT / res_name.stem.replace('beAHA','changed.csv'))
+            fname = list(data_io.LOCAL_OUTPUT.glob(f'pid={pid}*loc={loc_id}_beAHA*'))[0]
+            be_agout = read_agout(fname)
+            # Get the most cost effective strategy - before (be)
+            be_most_ce = _get_most_ce_strategy(be_agout)
+
+            # ORIGINAL CODE
+            # Note: not working because be site name won't appear in af_agout strategy columns
+            # Might happen because hospital treatment time is so long it passes the limit a stroke
+            # patient can get treatment
+            # af_qaly_stats = get_variable_stats_from_aggregated_outcome(af_agout,'QALY',
+            #                    [af_most_ce,be_most_ce.replace('- most C/E','').strip()],
+            #                    suffixes=("af","be"))
+
+            # Note: This chunk runs but be columns are all empty
+            # if be_most_ce not in af_agout.columns.get_level_values(0):
+            #     af_qaly_stats = get_variable_stats_from_aggregated_outcome(af_agout,'QALY',
+            #                     strategies=[af_most_ce],
+            #                     suffixes=["af"])
+
+            # If strategy name not in af_agout columns, only summarize af QALYs
+            if be_most_ce.replace('- most C/E','').strip() in af_agout.columns.get_level_values(0):
+                af_qaly_stats = get_variable_stats_from_aggregated_outcome(af_agout,'QALY',
+                                [af_most_ce,be_most_ce.replace('- most C/E','').strip()],
+                                suffixes=["af","be"])
+            else:
+                print("Missing Strategy: ", be_most_ce)
+                print("for location ", loc_id, "and patient ", pid)
+                af_qaly_stats = get_variable_stats_from_aggregated_outcome(af_agout,'QALY',
+                                [af_most_ce],
+                                suffixes=["af"])
+            
+            af_qaly_stats.index=[loc_id] # a dataframe
+            # fname = list(data_io.LOCAL_OUTPUT.glob(f'pid={pid}*loc={loc_id}*beAHA*'))[0]
+            # be_qaly_stats = get_variable_stats_from_aggregated_outcome(fname,'QALY')
+            # be_qaly_stats.index=[loc_id]
+            # qaly_stats = be_qaly_stats.join(af_qaly_stats,lsuffix='_be',rsuffix='_af')
+            qaly_stats_l.append(af_qaly_stats)
+
+        changed_res = changed_res.join(pd.concat(qaly_stats_l),how='left')
+    
+        # Create column for difference in mean QALYs
+        # mean_QALY_af - mean_QALY_be
+        changed_res['diff_QALY'] = changed_res['mean_QALY_af'] - changed_res['mean_QALY_be']
+
+        changed_res.to_csv(
+            data_io.BASIC_ANALYSIS_OUTPUT / res_name.stem.replace('beAHA','changed.csv'))
+
+
+print("Code took", time.time() - start_time, "to run")
